+++ src/main/java/com/yanny/ytech/registration/YTechItemTags.java
@@ -41,6 +41,8 @@ public class YTechItemTags {
     public static final TagKey<Item> SHARP_FLINTS = create("sharp_flints");
     public static final TagKey<Item> UNFIRED_BRICKS = create("unfired_bricks");
     public static final TagKey<Item> UNFIRED_CLAY_BUCKETS = create("unfired_clay_buckets");
+    public static final TagKey<Item> UNFIRED_DECORATED_POTS = create("unfired_decorated_pots");
+    public static final TagKey<Item> UNFIRED_FLOWER_POTS = create("unfired_flower_pots");
     public static final TagKey<Item> VENISON = create("venison");
     public static final TagKey<Item> WATER_BUCKETS = create("water_buckets");