+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Added Potter's Wheel
 - Changed Unfired Clay Bucket recipe
+- Added unfired versions for Flower Pot and Decorated Pot
 
 ## [0.3.3]